Fix nvidia_p2p_get_pages length alignment
The length of RDMA buffers needs to be aligned to 64Kb pages.
This has always been the case but the NVIDIA driver did not
previously check for this alignment. The 560 NVIDIA drivers
added this alignment check, so the caller now need to
make sure the length is aligned to avoid errors.